Fix compile on MSVC 2012 from Peter Ferrie (nw)

+++ src/mame/drivers/segaufo.c
@@ -372,7 +372,7 @@ void ufo_state::machine_start()
 {
 	// init/zerofill
 	static const float motor_speeds[4] =
-		{ 1.0/CABINET_WIDTH, 1.0/CABINET_DEPTH, 1.0/CABINET_HEIGHT, 1.0/CRANE_SIZE };
+		{ (float) (1.0/CABINET_WIDTH), (float) (1.0/CABINET_DEPTH), (float) (1.0/CABINET_HEIGHT), (float) (1.0/CRANE_SIZE) };
 	
 	for (int p = 0; p < 2; p++)
 	{
